Backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==4.2.7
django-auditlog==2.3.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.5.0
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
